remove autoprefixer dependency: make it optional
This change makes assetgraph use whatever autoprefixer version
that is installed with the project it is ran in. This is desireable
as you will otherwise potentially get different output when using
your projects version and the one in assetgraph.

We had this problem when upgrading autoprefixer from 1.x to 2.1
as we use a version of autoprefixer in our project for a some
middleware in our development server.

This is a breaking change in the sense that you are now required
to provide a version of autoprefixer for this transform to work.
It bails out gracefully, like the svg and less transforms do,
warning the user if cssAssets is found but no autoprefixer
was available.