buildProduction: Split the md5 renaming up into two steps
First we move the assets to the right directory, then (in a separate
pass) we add the hash to the file name.

Makes the hashes come out right when a moved asset has outgoing
relations whose hrefs change as a result of moving it. In pass two those
outgoing hrefs won't change now.