Use Node.js v14 for Docker image, not the latest one (v15)
This fixes building of the Docker image, which otherwise fails because the
node-canvas dependency doesn't have a prebuilt version available for Node.js
v15, and the image doesn't have the necessary system-level dependencies to
build it from source.

Using Node.js v14 (the latest LTS version) fixes this as we then can use the
prebuilt node-canvas binary.

As a bonus I have also added clearing of the NPM cache after installing
packages, this will decrease the final image size slightly.